Changed Elo to Double

+++ app/src/main/java/com/example/androidchess/User.kt
@@ -2,5 +2,5 @@ package com.example.androidchess
 
  object User{
      var name: String =""
-     var elo: Int = 1000
+     var elo: Double = 1000.0
 }
